chore: extract schema assert to its own file
